% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMARC.R
\docType{class}
\name{customCBARuleModel-class}
\alias{customCBARuleModel-class}
\alias{customCBARuleModel}
\title{rCBARuleModel}
\description{
This class represents an CBA rule-based classifier, where rules are represented as string vectors in a data frame
}
\section{Slots}{

\describe{
\item{\code{rules}}{dataframe output by \pkg{rCBA}}

\item{\code{cutp}}{list of cutpoints}

\item{\code{classAtt}}{name of the target class attribute}

\item{\code{attTypes}}{attribute types}
}}

