% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMARC.R
\name{predict.qCBARuleModel}
\alias{predict.qCBARuleModel}
\title{Aplies qCBARuleModel}
\usage{
\method{predict}{qCBARuleModel}(object, newdata, testingType,
  loglevel = "INFO", ...)
}
\arguments{
\item{object}{\link{qCBARuleModel} class instance}

\item{newdata}{data frame with data}

\item{testingType}{either "mixture" for multi rule classification or "firstRule" for one rule classification. Applicable only when model is loaded from file.}

\item{loglevel}{logger level from java.util.logging}

\item{...}{other arguments (currently not used)}
}
\value{
vector with predictions.
}
\description{
Method that matches qCBA rule model. Supports both one rule and multi rule classification.
}
\examples{
allData <- datasets::iris[sample(nrow(datasets::iris)),]
trainFold <- allData[1:100,]
testFold <- allData[101:nrow(datasets::iris),]
rmCBA <- cba(trainFold, classAtt="Species")
rmqCBA <- qcbaExtend(cbaRuleModel=rmCBA,datadf=trainFold)
print(rmqCBA@rules)
prediction <- predict(rmqCBA,testFold)
acc <- CBARuleModelAccuracy(prediction, testFold[[rmqCBA@classAtt]])
message(acc)

}
\seealso{
\link{qcbaExtend}
}
