% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMARC.R
\name{qcba}
\alias{qcba}
\title{qCBA Quantitative CBA}
\usage{
qcba(cbaRuleModel, datadf, continuousPruning = FALSE, postpruning = TRUE,
  fuzzification = FALSE, annotate = FALSE, ruleOutputPath,
  minImprovement = 0, minCondImprovement = -0.05, minConf = 0.5,
  extensionStrategy = "ConfImprovementAgainstLastConfirmedExtension",
  loglevel = "WARNING", createHistorySlot = FALSE)
}
\arguments{
\item{cbaRuleModel}{a \link{CBARuleModel}}

\item{datadf}{data frame with training data}

\item{continuousPruning}{indicating  if continuous pruning is enabled}

\item{postpruning}{boolean indicating if postpruning is enabled}

\item{fuzzification}{boolean indicating if fuzzification is enabled. Multi rule classification model is produced if enabled. Fuzzification without annotation is not supported.}

\item{annotate}{boolean indicating if annotation with probability distributions is enabled, multi rule classification model is produced if enabled}

\item{ruleOutputPath}{path of file to which model will be saved. Must be set if multi rule classification is produced.}

\item{minImprovement}{parameter ofqCBA extend procedure  (used when  extensionStrategy=ConfImprovementAgainstLastConfirmedExtension or ConfImprovementAgainstSeedRule)}

\item{minCondImprovement}{parameter ofqCBA extend procedure}

\item{minConf}{minimum confidence  to accept extension (used when  extensionStrategy=MinConf)}

\item{extensionStrategy}{possible values: ConfImprovementAgainstLastConfirmedExtension, ConfImprovementAgainstSeedRule,MinConf}

\item{loglevel}{logger level from java.util.logging}
}
\value{
Object of class \link{qCBARuleModel}.
}
\description{
Creates qCBA one rule or multi rule classication model from a CBA rule model
}
\examples{
allData <- datasets::iris[sample(nrow(datasets::iris)),]
trainFold <- allData[1:100,]
testFold <- allData[101:nrow(datasets::iris),]
rmCBA <- cba(trainFold, classAtt="Species")
rmqCBA <- qcba(cbaRuleModel=rmCBA,datadf=trainFold)
print(rmqCBA@rules)
}
