% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMARC.R
\name{qcbaExtend}
\alias{qcbaExtend}
\title{qCBA Extension with pruning options}
\usage{
qcbaExtend(cbaRuleModel, datadf, continuousPruning = FALSE,
  postpruning = TRUE, fuzzification = FALSE, annotate = FALSE,
  ruleOutputPath, loglevel = "FINEST")
}
\arguments{
\item{cbaRuleModel}{a \link{CBARuleModel}}

\item{datadf}{data frame with training data}

\item{continuousPruning}{indicating  if continuous pruning is enabled}

\item{postpruning}{boolean indicating if postpruning is enabled}

\item{fuzzification}{boolean indicating if fuzzification is enabled. Multi rule classification model is produced if enabled. Fuzzification without annotation is not supported.}

\item{annotate}{boolean indicating if annotation with probability distributions is enabled, multi rule classification model is produced if enabled}

\item{ruleOutputPath}{path of file to which model will be saved. Must be set if multi rule classification is produced.}

\item{loglevel}{logger level from java.util.logging}
}
\value{
Object of class \link{qCBARuleModel}.
}
\description{
Creates qCBA one rule or multi rule classication model from a CBA rule model
}
\examples{
allData <- datasets::iris[sample(nrow(datasets::iris)),]
trainFold <- allData[1:100,]
testFold <- allData[101:nrow(datasets::iris),]
rmCBA <- cba(trainFold, classAtt="Species")
rmqCBA <- qcbaExtend(cbaRuleModel=rmCBA,datadf=trainFold)
print(rmqCBA@rules)
}
