% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMARC.R
\name{qcbaHumTemp}
\alias{qcbaHumTemp}
\title{Use the Humidity-Temperature toy dataset from the arc package to test one rule classification QCBA workflow.}
\usage{
qcbaHumTemp()
}
\value{
QCBA model
}
\description{
TODO
}
