% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMARC.R
\name{qcbaIris}
\alias{qcbaIris}
\title{Use the \link{iris} dataset to the test QCBA workflow.}
\usage{
qcbaIris()
}
\value{
Accuracy.
}
\description{
Learns a CBA classifier and performs all QCBA postprocessing steps
}
