% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMARC.R
\name{rcbaModel2CustomCBAModel}
\alias{rcbaModel2CustomCBAModel}
\title{rcbaModel2CustomCBAModel Converts a model created by \pkg{rCBA} so that it can be passed to qCBA}
\usage{
rcbaModel2CustomCBAModel(rcbaModel, cutPoints, classAtt, rawDataset, attTypes)
}
\arguments{
\item{rcbaModel}{object returned  by \link[rCBA]{build}}

\item{cutPoints}{specification of cutpoints applied on the data before they were passed to  \link[rCBA]{build}}

\item{classAtt}{the name of the class attribute}

\item{rawDataset}{the raw data (before discretization). This dataset is used to guess attribute types if attTypes is not passed}

\item{attTypes}{vector of attribute types of the original data.  If set to null, you need to pass rawDataset.}
}
\description{
Creates instance of \link{customCBARuleModel} class based on model created by the  \pkg{rCBA}  package.
Instance of \link{customCBARuleModel} can then be passed to \link[qCBA]{qcba} instead of an instance of \link{CBARuleModel},
which is created with the arc package.
}
\examples{
if (! requireNamespace("rCBA", quietly = TRUE)) {
 message("Please install rCBA: install.packages('rCBA')")
} else {
 discrModel <- discrNumeric(iris, "Species")
 irisDisc <- as.data.frame(lapply(discrModel$Disc.data, as.factor))
 rCBAmodel <- rCBA::build(irisDisc)
 cCBAmodel <- rcbaModel2CustomCBAModel(rCBAmodel,discrModel$cutp,"Species",iris)
 qCBAmodel <- qcba(cCBAmodel,iris)
 print(qCBAmodel@rules)
}

}
