% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMARC.R
\name{mapDataTypes}
\alias{mapDataTypes}
\title{Map R types to qCBA}
\usage{
mapDataTypes(Rtypes)
}
\arguments{
\item{Rtypes}{Vector with R data types}
}
\value{
Vector with qCBA data types
}
\description{
The QCBA Java implementation uses different names of some data types than are used in this R wrapper.
}
\examples{
mapDataTypes(unname(sapply(iris, class)))
}
