% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMARC.R
\name{qcbaHumTemp}
\alias{qcbaHumTemp}
\title{Use the HumTemp dataset to test the one rule classification QCBA workflow.}
\usage{
qcbaHumTemp()
}
\value{
QCBA model
}
\description{
Learns a CBA classifier and performs all QCBA postprocessing steps.
}
