% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMARC.R
\name{qcbaIris}
\alias{qcbaIris}
\title{Use the Iris dataset to test one rule classification QCBA workflow.}
\usage{
qcbaIris()
}
\value{
Accuracy.
}
\description{
Learns a CBA classifier, performs all qCBA postprocessing steps for one rule classification.
}
