% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMARC.R
\name{qcbaIris}
\alias{qcbaIris}
\title{Use the Iris dataset to test to test one rule classification qCBA workflow.}
\usage{
qcbaIris()
}
\value{
Accuracy.
}
\description{
Learns a CBA classifier, performs qCBA Extension with continuous pruning and postpruning. Applies the model in one rule classification.
}
