% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMARC.R
\name{qcbaIris}
\alias{qcbaIris}
\title{Use the \link{iris} dataset to the test QCBA workflow.}
\usage{
qcbaIris()
}
\value{
Accuracy.
}
\description{
Learns a CBA classifier and performs all QCBA postprocessing steps
This function uses a package-specific seed option, which can be overridden 
with `options(qcba.seed = 42)`, where 42 is an example seed value
}
