% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMARC.R
\name{qcbaIris2}
\alias{qcbaIris2}
\title{Use the Iris dataset to test multi rule qCBA workflow.}
\usage{
qcbaIris2()
}
\value{
Accuracy.
}
\description{
Learns a CBA classifier, performs qCBA with the experimental multi rule classification workflow including annotation and fuzzification. Applies the model with rule mixture classification.
The model  is saved to a temporary file.
}
