% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMARC.R
\name{qcbaIris2}
\alias{qcbaIris2}
\title{Use the Iris dataset to test the experimental multi-rule QCBA workflow.}
\usage{
qcbaIris2()
}
\value{
Accuracy.
}
\description{
Learns a CBA classifier, and then transforms it  to a multirule classifier,
 including rule annotation and fuzzification. Applies the learnt model with rule mixture classification.
The model  is saved to a temporary file.
}
